@@ -6,13 +6,14 @@ repository: https://github.com/Flutter-Bounty-Hunters/follow_the_leader
 version: 0.0.4+8
 
 environment:
-  sdk: ">=2.15.0 <4.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   logging: ^1.0.1
+  super_keyboard: ^0.2.2
   vector_math: ^2.1.2
 
 dev_dependencies:
@@ -20,4 +21,7 @@ dev_dependencies:
     sdk: flutter
   flutter_lints: ^1.0.0
 
+  flutter_test_goldens: 0.0.7
+
 flutter:
+
