@@ -83,10 +83,7 @@ class _OrbitingCirclesDemoState extends State<OrbitingCirclesDemo> {
           radius * cos(radians),
           radius * sin(radians),
         ),
-        boundary: ScreenFollowerBoundary(
-          screenSize: MediaQuery.of(context).size,
-          devicePixelRatio: MediaQuery.of(context).devicePixelRatio,
-        ),
+        boundary: const ScreenFollowerBoundary(),
         leaderAnchor: Alignment.center,
         followerAnchor: Alignment.center,
         child: Container(
