@@ -1,31 +1,33 @@
 // ignore_for_file: avoid_print
 
-import 'package:logging/logging.dart' as logging;
+import 'package:logging/logging.dart';
 
 /// Follow the Leader logger names.
-class LogNames {
-  static const leader = 'leader';
-  static const follower = 'follower';
-  static const link = 'link';
-  static const boundary = 'boundary';
-  static const widgetBoundary = 'boundary.widget';
+abstract class FtlLogNames {
+  static const leader = 'ftl.leader';
+  static const follower = 'ftl.follower';
+  static const link = 'ftl.link';
+  static const boundary = 'ftl.boundary';
+  static const widgetBoundary = 'ftl.boundary.widget';
+  static const _root = 'ftl';
 }
 
 /// Follow the Leader logging.
-class FtlLogs {
-  static final leader = logging.Logger(LogNames.leader);
-  static final follower = logging.Logger(LogNames.follower);
-  static final link = logging.Logger(LogNames.link);
-  static final boundary = logging.Logger(LogNames.boundary);
-  static final widgetBoundary = logging.Logger(LogNames.widgetBoundary);
+abstract class FtlLogs {
+  static final leader = Logger(FtlLogNames.leader);
+  static final follower = Logger(FtlLogNames.follower);
+  static final link = Logger(FtlLogNames.link);
+  static final boundary = Logger(FtlLogNames.boundary);
+  static final widgetBoundary = Logger(FtlLogNames.widgetBoundary);
+  static final _root = Logger(FtlLogNames._root);
 
-  static final _activeLoggers = <logging.Logger>{};
+  static final _activeLoggers = <Logger>{};
 
   /// Initialize the given [loggers] using the minimum [level].
   ///
   /// To enable all the loggers, use [FtlLogs.initAllLogs].
-  static void initLoggers(logging.Level level, Set<logging.Logger> loggers) {
-    logging.hierarchicalLoggingEnabled = true;
+  static void initLoggers(Set<Logger> loggers, [Level level = Level.ALL]) {
+    hierarchicalLoggingEnabled = true;
 
     for (final logger in loggers) {
       if (!_activeLoggers.contains(logger)) {
@@ -42,8 +44,8 @@ class FtlLogs {
   /// Initializes all the available loggers.
   ///
   /// To control which loggers are initialized, use [FtlLogs.initLoggers].
-  static void initAllLogs(logging.Level level) {
-    initLoggers(level, {logging.Logger.root});
+  static void initAllLogs([Level level = Level.ALL]) {
+    initLoggers({_root}, level);
   }
 
   /// Returns `true` if the given [logger] is currently logging, or
@@ -54,12 +56,12 @@ class FtlLogs {
   /// information that's costly to compute. In such a case, you can
   /// choose to compute the expensive information only if the given
   /// logger will actually log the information.
-  static bool isLogActive(logging.Logger logger) {
+  static bool isLogActive(Logger logger) {
     return _activeLoggers.contains(logger);
   }
 
   /// Deactivates the given [loggers].
-  static void deactivateLoggers(Set<logging.Logger> loggers) {
+  static void deactivateLoggers(Set<Logger> loggers) {
     for (final logger in loggers) {
       if (_activeLoggers.contains(logger)) {
         print('Deactivating logger: ${logger.name}');
@@ -71,7 +73,7 @@ class FtlLogs {
   }
 
   /// Logs a record using a print statement.
-  static void _printLog(logging.LogRecord record) {
+  static void _printLog(LogRecord record) {
     print(
         '(${record.time.second}.${record.time.millisecond.toString().padLeft(3, '0')}) ${record.loggerName} > ${record.level.name}: ${record.message}');
   }
